requirements.txt
matplotlib
scipy
opencv-python
keras
numpy
flask
flask-cors
tensorflow
pygame
pyrebase4